Moved all TICA tests in to test_tica.py
